Refactor to a Unit behaviour
Count and Duration implement the Unit behaviour
